vcard profile: simplified REV workaround
The scripting code seemed a bit overly complex. There shouldn't be
a need for the temporary variable.

The original patch also had the side-effect of always setting
REV to its normalized version, even if it wasn't empty. That
may or may not influence other syncs, but it is better to not
take the risk.