syncevo-dbus-server: fixed a compile problem with KWallet + libnotify
When both KWallet and libnotify support were enabled, the redefinition
of "signals" in the Qt headers broke the GTK headers that are pulled in
by notify.h, because one of the GTK structs has a "signals" member.

Fixed by asking Qt to not define these special keywords as
preprocessor symbols.