KWallet: do not enable by default
This is motivated by the observation that KWallet was unintentionally
enabled just because the header file had been installed. It then
also failed to compile because of missing Qt compile flags (separate
problem), but even if it had successed, the extra library dependency
wasn't wanted in that case, so better request an explicit --enable
before enabling the feature.