vCard profile: added more permanent support for KAddressBook extensions
This re-adds the code from a3d74129726c0de0164d00b7fe0f904904206615
which was temporarily removed in revert commit
4ff3777c705c7c6904feed93a8cac4fab2295fd8.

This time, the changes are done in such a way that other, non-KDE
backends are not affected. The properties fall into three categories:

1. The KDE extension maps to a field which has already an property mapping
   that is understood by SyncML peers (example: X-messaging/xmpp-All KDE
   property maps to X-JABBER). In these cases, the external
   property is not used with KDE (empty property with rule="KDE").
   A second property entry exclusively for KDE (rule="KDE") defines
   the mapping to the KDE property.
2. The KDE extension has no corresponding mapping (X-KADDRESSBOOK-CRYPTO*,
   X-messaging/irc-All for IRC chat handle). In these cases the KDE
   property name is used also with SyncML peers.
3. An Evolution extension has no corresponding extension in KDE (X-EVOLUTION-VIDEO-URL).
   Here the Evolution extension is passed to KDE.

Whether KDE extensions like X-messaging/irc-All should be exposed as
they are to other peers is uncertain. It might be more natural to follow
the X-AIM/JABBER/MSM/... pattern and introduce X-IRC. Or follow RFC 4770,
which defines the IMPP property... at the end of the day, this depends on
the peers which we want to send these extensions to.