WebDAV: send all optional extensions to CalDAV/CardDAV servers
Some of our extensions are optional, for example X-EVOLUTION-UI-SLOT
(because Nokia phones had problems with that). CalDAV/CardDAV servers
are meant to preserve all extensions, so let's send them everything we
have.

For calendar, this is not yet relevant in practice (rule doesn't do
anything).

Apple Calendar server + CardDAV really preserves the slot parameter.
Yahoo CardDAV does not, but at least accepts it.