client-test: disabled reversed combination of sources
The combination of all enabled sources was tested twice,
once in the order specified by CLIENT_TEST_SOURCES,
once reversed.

This used to make a difference with the Funambol SyncML
library, but not anymore with Synthesis. Disabled this
because of the increased time for testing. Can be
enabled again by setting CLIENT_TEST_REVERSE_SOURCES.