WebDAV: ignore 404 for DELETE requests
This could be caused by a race condition (item already deleted by some
other client). In practice it was seen with a confused Google Calendar
server: event listed by REPORT, but both GET and DELETE fail for
it. Not much that can be done about that event via CalDAV then, so let's
ignore it.