TrackingSyncSource: allow intercepting SyncSource callbacks
The main usage is for derived classes which cannot do all of their
initialization in open() because it is too heavy. The implementation
of the D-Bus CheckSource() call expects open() to be fast and local.
The more expensive initialization needs to be done in beginSync(),
which requires calling TrackingSyncSource::beginSync() when
done.