local sync: better calendar slow sync (partly fixes BMC #14804)
This patch replaces the ugly configuration translation with a runtime
check in a comparescript.

This is a first step towards detecting properly at runtime whether a
peer supports UID/RECURRENCE-ID semantic in calendar data. Currently
this check is still based on the "local sync == use UID/RECURRENCE-ID"
shortcut.

This patch depends on a libsynthesis which supports the COMPAREMODE()
method.