Google CalDAV: fixed corner case in update hack
When updating a meeting series by adding a child, the series' SEQ
value was not increased if the new child already had a higher SEQ
value. Then the updating failed because the other events in the series
didn't have a higher SEQ than on the server.

Fix this by always bumping first, then comparing against the new
event.