Google CalDAV: fixed corner cases in delete hack
There were still several cases where the "Google delete hack" (= catch
 409 "Can't delete a recurring event except on its organizer's
 calendar", update item, delete again) did not work.

Sometimes it failed when only an EXDATE was set => also remove
EXDATE, not just RRULE, to convince the server that the event is not
recurring.

Sometimes it failed because the DELETE came too soon after the PUT
(?!). Added a retry loop.