Evolution: enable preservation of local data (BMC #15030)
Local data not supported by a peer (for example, X- extensions and
vCard 3.0 properties) was lost when importing updates from such a peer.
The Synthesis engine can preserve such extensions, but doesn't
apply that more expensive merging unless a backend is configured
with <updateallfields>.

This patch adds it for the Evolution contact and calendar backends.