Google testing: simplify test cases instead of ignoring differences in synccompare
The motivation for this is that the test cases are also used in the
new testExtensions test. Because that test runs with the "google"
hacks in synccompare disabled, it breaks for those test cases which
contain properties that get mangled by Google. This includes
everything with a semicolon in it where Google doesn't expect it
(NOTE, ORG).