client-test: make testExtensions more permissive
testExtensions compares data without per-peer hacks. This is
problematic for properties which are supported, but get mangled by a
peer.

This patch addresses the problem by disabling the checking of
properties and parameters which are expected to be problematic for all
peers: PHOTO (might get transcoded), FN (often synthesized) and
X-EVOLUTION-UI-SLOT (preserving parameters problematic).

It would be better to make this supression configurable per-peer
again, but that's overkill at the moment.