WebDAV: moved common href and etag handling into common parser
The handling of href and etag data had to be done by all
classes using initResponseHandler(). Now it is in the XML
handler itself.

In addition, users of the class process the gathered results
at the end of the results XML element. At that point, all
the response parts are guaranteed to be processed.