testing: avoid race condition in testLinkedItem* tests
The testLinkedItem tests updated items without actually
modifying their content. That worked as long as LAST-MODIFIED
increased, but because of the recent speed-ups that was no
longer always the case. The test then failed because the ETag
of a CalDAV resource didn't changed on the Apple Calendar Server.

This commit fixes that by modifying the item content when it is
to be used as an update.

One test didn't take that into account and continued using the
unmodified item content. Fixed.