CalDAV: ensure that UID is set during update
Updating an item must be done with the same UID that was originally
set by the server. The Maemo 5 backend replaces the UID received from
the server with its own sequential numbering of events in the SQLite
database.

This commit is an attempt to catch this situation and restore the
correct the UID before sending the update item content to the server.