CalDAV: handle response with no data (Google)
Moving the response handling from the data element to the response
element caused problems with Google, because it sends a 404 status for
the collection with no data. Apple Calendar Server didn't do that when
testing the change manually, so the problem only showed up in the
nightly testing.

This patch restores the previous behavior of simply ignoring responses
with no data. Some better error handling might be useful.